pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ucgen"
version = "0.1.0"
description = "Generate structured use case documents from natural language input."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "typer[all]>=0.12",
  "httpx>=0.27",
  "anthropic>=0.30",
  "openai>=1.40",
  "pydantic>=2.7",
  "jinja2>=3.1",
  "rich>=13.7",
  "aiofiles>=23.0",
]

[project.optional-dependencies]
project = ["PyYAML>=6.0"]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "pytest-mock>=3.12",
  "ruff>=0.5",
  "mypy>=1.10",
  "types-PyYAML>=6.0",
]

[project.scripts]
ucgen = "ucgen.cli:app"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = []

[tool.ruff.format]
quote-style = "double"
